# api/openapi.yaml
openapi: 3.0.3
info:
  title: Village Calendar Service REST API
  version: 1.0.0
  description: REST API for webhooks, OAuth2 authentication, and health monitoring. For calendar operations, use the GraphQL API at /graphql.
  contact:
    name: VillageCompute Support
    url: https://villagecompute.com
servers:
  - url: http://localhost:8030
    description: Development server
  - url: https://calendar.villagecompute.com
    description: Production server
...
paths:
  /api/webhooks/stripe:
    post:
      tags:
        - Webhooks
      summary: Handle Stripe webhook events
      description: |
        Processes Stripe payment webhook events with signature verification.
        This endpoint is called by Stripe when payment events occur.
        Supported event types: payment_intent.succeeded, payment_intent.payment_failed.
      operationId: handleStripeWebhook
      parameters:
        - name: Stripe-Signature
          in: header
          required: true
          description: |
            HMAC-SHA256 signature generated by Stripe using the webhook secret.
            Required for webhook signature verification to prevent replay attacks.
          schema:
            type: string
          example: "t=1234567890,v1=abc123..."
      requestBody:
        required: true
        description: Raw JSON payload from Stripe webhook event
        content:
          application/json:
            schema:
              type: string
            example: |
              {
                "id": "evt_...",
                "type": "payment_intent.succeeded",
                "data": {...}
              }
      responses:
        '200':
          description: Webhook processed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookSuccessResponse'
              example:
                status: success
        '400':
          description: Invalid signature or malformed payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookErrorResponse'
              examples:
                invalid_signature:
                  value:
                    error: "Invalid signature"
                missing_signature:
                  value:
                    error: "Missing signature header"
                invalid_payload:
                  value:
                    error: "Invalid webhook payload"
        '500':
          description: Internal server error processing webhook
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookErrorResponse'
              example:
                error: "Failed to process event"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from OAuth2 callback endpoints
  schemas:
    WebhookSuccessResponse:
      type: object
      description: Successful webhook processing response
      properties:
        status:
          type: string
          description: Status message
          example: success
    WebhookErrorResponse:
      type: object
      description: Webhook error response
      properties:
        error:
          type: string
          description: Error message describing what went wrong
          example: "Invalid signature"
...